a2a/a2a-lambda-python/src/server/requirements.txt
boto3
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
strands-agents[a2a]
uvicorn